spec: added RECV_BUSY error and moved state modifications.
That is, we only modify the TCU state after a SEND/REPLY has been
executed successfully. The state is in this case the credits and the
unread/occupied mask. Performing these changes only after a success
allows the sender to retry the command in case he received a RECV_BUSY
error.