spec,python,nocrw: added TCU version and pass it through.
The TCU has now a version number integrated into the FEATURES register
of the external interface. M³ is built for a specific TCU version,
depending on the target (hw22=0, hw=1) and passes this version to the
python interface. Before loading software onto the FPGA, M³ loads the
TCU version from all TCUs and checks whether the version matches.

In python, the TCU class was re-implemented to not contain class
members, but functions. These functions are called from other modules
and internally use the version number to return the correct address for
the corresponding TCU version.

Additionally, the TCU version is passed to nocrw, which needs it to send
messages with the correct message header, depending on the version, to
the FPGA.